rearrange scripts for reusability
