+++ packs/proc/proc.py
@@ -3,7 +3,7 @@ import os
 from packs.core.io                import read_config_file
 from packs.proc.processing_utils  import process_bin_WD2
 from packs.proc.processing_utils  import process_bin_WD1
-from packs.proc.waveform_utils    import calibrate
+from packs.proc.calibration_utils    import calibrate
 from packs.core.core_utils        import check_test
 
 def proc(config_file):
